requirements.txt
graphene-django
django-filter
Django